+++ spec/requests/host_meta_request_spec.rb
@@ -1,0 +1,12 @@
+require "rails_helper"
+
+describe "The host_meta route" do
+  describe "requested without accepts headers" do
+    it "returns an xml response" do
+      get host_meta_url
+
+      expect(response).to have_http_status(:success)
+      expect(response.content_type).to eq "application/xrd+xml"
+    end
+  end
+end
