+++ spec/routing/well_known_routes_spec.rb
@@ -1,0 +1,15 @@
+require 'rails_helper'
+
+describe 'the host-meta route' do
+  it 'routes to correct place with xml format' do
+    expect(get('/.well-known/host-meta')).
+      to route_to('well_known/host_meta#show', format: 'xml')
+  end
+end
+
+describe 'the webfinger route' do
+  it 'routes to correct place with json format' do
+    expect(get('/.well-known/webfinger')).
+      to route_to('well_known/webfinger#show', format: 'json')
+  end
+end
